+++ frontend/src/Settings.css
@@ -1,47 +1,103 @@
 :root{
-    --bg:#2f3640; --bg-deeper:#2b313a; --panel:#3a414c;
-    --txt:#e8edf3; --muted:#9aa3ad; --accent:#1c7e91; --danger:#d35b5b;
+    --bg:#2f3640;
+    --bg-deeper:#2b313a;
+    --panel:#3a414c;
+    --txt:#e8edf3;
+    --muted:#9aa3ad;
+    --accent:#1c7e91;
+    --danger:#d35b5b;
     --ring:#4a5666;
+    --shadow: rgba(0,0,0,.35);
 }
 
-.st-wrap{display:flex; height:100vh; background:var(--bg); color:var(--txt);}
+/* 전체 화면 레이아웃 */
+.settings-page {
+    display:flex;
+    height:100vh;
+    width:100vw;
+    background: var(--bg);
+    color: var(--txt);
+    overflow:hidden;
+}
 
-.st-aside{width:240px; background:var(--bg-deeper); padding:20px 16px; box-shadow: inset -1px 0 0 #00000022;}
-.st-logo{font-weight:800; letter-spacing:1px; margin-bottom:16px;}
-.st-sec-title{font-size:12px; color:var(--muted); margin-top:12px; margin-bottom:6px;}
-.mt24{margin-top:24px;}
+/* 왼쪽 사이드바 */
+.settings-sidebar {
+    width: 260px;
+    background: var(--bg-deeper);
+    padding: 20px 14px;
+    box-shadow: inset -1px 0 0 #00000022;
+    display:flex;
+    flex-direction: column;
+    gap: 6px;
+}
+.settings-logo {
+    font-weight: 800;
+    font-size: 20px;
+    margin-bottom: 16px;
+}
+.sidebar-section-title {
+    font-size: 12px;
+    color: var(--muted);
+    margin: 8px 6px 4px;
+}
+.sidebar-divider {
+    height: 1px;
+    background: #00000033;
+    margin: 10px 4px;
+    border-radius: 1px;
+}
+.sidebar-item {
+    appearance: none;
+    border: 0;
+    background: transparent;
+    color: var(--txt);
+    text-align: left;
+    padding: 10px 12px;
+    border-radius: 8px;
+    opacity: .9;
+    cursor: pointer;
+}
+.sidebar-item:hover {
+    background: #ffffff12;
+    opacity: 1;
+}
+.sidebar-item.is-active {
+    background: #ffffff18;
+    font-weight: 600;
+}
 
-.st-nav{display:flex; flex-direction:column; gap:6px; margin-bottom:8px;}
-.st-nav-item{padding:10px 12px; border-radius:8px; color:var(--txt); text-decoration:none; opacity:.85;}
-.st-nav-item:hover{background:#ffffff10;}
-.st-nav-item.active{background:#ffffff18; font-weight:600;}
+/* 오른쪽 본문 */
+.settings-main {
+    flex:1;
+    display:flex;
+    flex-direction: column;
+    overflow:hidden;
+}
+.settings-header {
+    padding: 20px 28px;
+    border-bottom: 1px solid #00000022;
+}
+.settings-h1 {
+    font-size: 22px;
+    font-weight: 700;
+    margin:0;
+}
+.settings-panel {
+    flex:1;
+    padding: 24px 28px;
+    overflow:auto;
+}
 
-.st-main{flex:1; padding:28px;}
-
-.st-panel{max-width:720px;}
-.st-h1{font-size:22px; font-weight:700; margin-bottom:16px;}
-
-.st-form{display:grid; grid-template-columns:1fr; gap:12px; max-width:420px;}
-.st-label{font-size:12px; color:var(--muted);}
-.st-input{background:var(--panel); border:1px solid transparent; color:var(--txt);
-    border-radius:10px; padding:10px 12px; outline:none;}
-.st-input:focus{border-color:var(--ring);}
-
-.st-pwbox{position:relative;}
-.st-input.pw{padding-right:44px;}
-.st-eye{position:absolute; top:50%; right:6px; transform:translateY(-50%);
-    width:32px; height:32px; border:none; border-radius:8px; background:#ffffff12; color:#fff; cursor:pointer;}
-.st-eye:hover{background:#ffffff20;}
-
-.st-primary{margin-top:8px; background:var(--accent); color:#fff; border:none; padding:10px 14px;
-    border-radius:8px; cursor:pointer; font-weight:600;}
-.st-primary:hover{filter:brightness(1.05);}
-
-.st-card{border-radius:12px; padding:16px; max-width:560px;}
-.st-h3{font-size:16px; margin-bottom:8px;}
-.st-ul{margin:8px 0 12px 0; padding-left:18px; color:#d5dbe3;}
-.st-row{display:flex; gap:8px;}
-.st-btn{background:#ffffff12; color:#fff; border:none; padding:10px 12px; border-radius:8px; cursor:pointer;}
-.st-btn:hover{background:#ffffff20;}
-.st-btn.danger{background:var(--danger);}
-.st-btn.danger:hover{filter:brightness(1.05);}
+/* 패널 내부 (MyAccount 등) */
+.st-form { display:grid; gap:12px; max-width:420px; }
+.st-label { font-size:12px; color:var(--muted); }
+.st-input {
+    background:var(--panel); border:1px solid transparent; color:var(--txt);
+    border-radius:10px; padding:10px 12px; outline:none;
+}
+.st-input:focus { border-color:var(--ring); }
+.st-primary {
+    background:var(--accent); color:#fff; border:none; padding:10px 14px;
+    border-radius:8px; cursor:pointer; font-weight:600;
+}
+.st-primary:hover{ filter:brightness(1.05); }
